// src/app/models/article.model.ts
export class Article {
  artikel_nr: number;
  bezeichnung: string;
  h_article_nr: number;
  bestell_nr: number;
  ek_netto: number;
  rabatt: number;
  shop_vk: number;
  ebay_vk: number;
  amazon_vk: number;
  kapazitat: number;
  farbe: string;
  gewicht: number;
  hersteller: string;
  status: string;
}